requirements.txt
asgiref==3.5.2
astroid==2.12.13
attrs==22.1.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.4
defusedxml==0.7.1
dill==0.3.6
Django==3.2
django-stubs==1.13.1
django-stubs-ext==0.7.0
django-templated-mail==1.1.1
djangorestframework==3.12.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg2==1.19.4
exceptiongroup==1.0.4
flake8==6.0.0
flake8-bugbear==22.12.6
flake8-builtins==2.0.1
flake8-import-order==0.18.2
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.3
pep8-naming==0.13.2
platformdirs==2.6.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pylint==2.15.8
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.14.1
